update imports
had to update config & add some docstring function (the latter work but need to be rewritten to be properly pythonic)

adds helper functions for each import file, which draws from the module docstring. The import function in ipython.imports also draws from the module docstring of what it is going to import.